added WLAN/Battery function again, battery needs calibration process, gesides that tested ok

+++ shared/sensors/spbox_sensors.cpp
@@ -303,12 +303,13 @@ void SPBOX_SENSORS::getPressure(float * pressure) { *pressure = pressure_; }
 void SPBOX_SENSORS::getAltitude(float * altitude) { *altitude = altitude_; }
 
 void SPBOX_SENSORS::updateVBat() {
-	uint8_t vbatADC;
+	uint16_t vbatADC;
 	float vbatFloat = 0.0F;
 	float vbatLSB = 0.97751F;		// 1000mV/1023 -> mV per LSB
 	float vbatVoltDiv = 0.201321;	// 271K/1271K resistor voltage divider
 
 	vbatADC = analogRead(VBAT_PIN);
+
 	vbatFloat = ((float)vbatADC * vbatLSB) / vbatVoltDiv;
 	vbatFloat /= 1000.;
